Adds indicator of a player visible voting log
A new player can be suprised that this tool keeps a history of
nominations. This indicator shows a book with a skull on it
when there is a nomination log present in the session object
along with the count of nominations stored locally in the
session object.

This allows a new user to see that there is a voting log.

This allows groups that wish to only see a limited voting
log to remind the storyteller to clear the log if they see
that there is one.

If accepted, this would address #124 and the secondary
considerations metioned in #121.

This commit squashes away the previous textual version